pr 리뷰 내용 수정

+++ src/main/java/com/example/team8salecommerce/domain/cart/controller/CartController.java
@@ -6,6 +6,8 @@ import com.example.team8salecommerce.domain.cart.dto.request.UpdateCartItemReque
 import com.example.team8salecommerce.domain.cart.dto.response.CartItemResponse;
 import com.example.team8salecommerce.domain.cart.dto.response.CartResponse;
 import com.example.team8salecommerce.domain.cart.service.CartService;
+import com.example.team8salecommerce.global.exception.CustomException;
+import com.example.team8salecommerce.global.exception.ErrorCode;
 import com.example.team8salecommerce.global.response.ApiResponse;
 import com.example.team8salecommerce.global.security.AuthMember;
 import jakarta.validation.Valid;
@@ -27,14 +29,20 @@ public class CartController {
             @AuthenticationPrincipal AuthMember authMember,
             @Valid @RequestBody AddCartItemRequest request
     ) {
+        if (authMember == null) {
+            throw new CustomException(
+                    ErrorCode.UNAUTHORIZED
+            );
+        }
+        {
+            CartItemResponse response =
+                    cartService.addCartItem(
+                            authMember.memberId(),
+                            request
+                    );
 
-        CartItemResponse response =
-                cartService.addCartItem(
-                        authMember.memberId(),
-                        request
-                );
-
-        return ResponseEntity.ok(ApiResponse.success(response));
+            return ResponseEntity.ok(ApiResponse.success(response));
+        }
     }
 
     // 장바구니 조회
@@ -42,9 +50,16 @@ public class CartController {
     public ResponseEntity<ApiResponse<CartResponse>> getCart(
             @AuthenticationPrincipal AuthMember authMember
     ) {
-        CartResponse response =
-                cartService.getCart(authMember.memberId());
-        return ResponseEntity.ok(ApiResponse.success(response));
+        if (authMember == null) {
+            throw new CustomException(
+                    ErrorCode.UNAUTHORIZED
+            );
+        }
+        {
+            CartResponse response =
+                    cartService.getCart(authMember.memberId());
+            return ResponseEntity.ok(ApiResponse.success(response));
+        }
     }
 
     // 장바구니 상품 수량 변경
@@ -55,15 +70,22 @@ public class CartController {
             @PathVariable Long cartItemId,
             @Valid @RequestBody UpdateCartItemRequest request
     ) {
+        if (authMember == null) {
+            throw new CustomException(
+                    ErrorCode.UNAUTHORIZED
+            );
+        }
+        {
 
-        CartItemResponse response =
-                cartService.updateCartItemQuantity(
-                        authMember.memberId(),
-                        cartItemId,
-                        request
-                );
+            CartItemResponse response =
+                    cartService.updateCartItemQuantity(
+                            authMember.memberId(),
+                            cartItemId,
+                            request
+                    );
 
-        return ResponseEntity.ok(ApiResponse.success(response));
+            return ResponseEntity.ok(ApiResponse.success(response));
+        }
     }
 
     // 장바구니 상품 삭제
@@ -73,12 +95,19 @@ public class CartController {
             @AuthenticationPrincipal AuthMember authMember,
             @PathVariable Long cartItemId
     ) {
+        if (authMember == null) {
+            throw new CustomException(
+                    ErrorCode.UNAUTHORIZED
+            );
+        }
+        {
 
-        cartService.deleteCartItem(
-                authMember.memberId(),
-                cartItemId
-        );
+            cartService.deleteCartItem(
+                    authMember.memberId(),
+                    cartItemId
+            );
 
-        return ResponseEntity.ok(ApiResponse.success(null));
+            return ResponseEntity.ok(ApiResponse.success(null));
+        }
     }
 }
